pyproject.toml
[project]
name = "image-jepa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "toml>=0.10.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
